# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OxyGen CMS API
  version: 0.8.0
  summary: OxyGen Central Management Server API
  description: |
    API contract for OxyGen Central Management Server Phase 1.

    Current CMS schema version: 0.19.

    Phase 1 APIs cover first-run setup, local authentication/RBAC, tenant/user/group/role administration,
    manual OxyGen instance enrollment, dashboard background poller status/controls, application logs, grid preferences, application settings,
    Settings system visibility, and Phase 1.5 queue status. Protected routes use bearer authentication plus explicit permissions and Tenant/global/instance scope checks.
servers:
  - url: http://localhost:3000
    description: Local API development server
  - url: https://{host}
    description: Deployed CMS server
    variables:
      host:
        default: cms.example.com
security:
  - bearerAuth: []
...
paths:
...
  /api/instances:
    get:
      tags: [Instances]
      summary: List accessible OxyGen instances
      parameters:
        - in: query
          name: includeArchived
          schema: { type: boolean, default: false }
          description: Include archived instances in the response. Archived instances are hidden from the default instance list but retained for dashboards/history and can be unarchived.
      responses:
        '200':
          description: Instance list scoped to the signed-in user's access.
          content:
            application/json:
              schema:
                type: object
                required: [instances]
                properties:
                  instances:
                    type: array
                    items: { $ref: '#/components/schemas/OxyGenInstance' }
        '401': { $ref: '#/components/responses/Error' }
    post:
      tags: [Instances]
      summary: Enroll OxyGen instance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateInstanceInput' }
      responses:
        '201':
          description: Created instance.
          content:
            application/json:
              schema:
                type: object
                required: [instance]
                properties:
                  instance: { $ref: '#/components/schemas/OxyGenInstance' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/instances/export.csv:
    get:
      tags: [Instances]
      summary: Export OxyGen instances as CSV
      description: SystemAdmin and TenantAdmin. Global users receive all instances with a `tenant` column containing Tenant names; blank tenant means global/unassigned. Tenant-scoped users receive only their Tenant's instances and the `tenant` column is omitted. The CSV includes `check_license`, `archived`, `metadata`, and `notes`; password values are never exported and the password column is always blank.
      responses:
        '200':
          description: CSV file of visible instances.
          content:
            text/csv:
              schema:
                type: string
                description: CSV with `instance_guid`, connection/configuration fields, and a blank `password` column. Global exports include `tenant` by Tenant name; tenant-scoped exports omit it.
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/instances/import:
    post:
      tags: [Instances]
      summary: Import OxyGen instances from CSV
      description: SystemAdmin and TenantAdmin. Upserts by `instance_guid` (the instance `id`). Blank `instance_guid` creates a new instance; an unknown nonblank GUID creates an instance with that GUID. Global users may include a `tenant` column with Tenant names, leaving it blank for global/unassigned instances; missing Tenant names are created for new instance rows. Tenant-scoped users must omit the `tenant` column; all creates/updates are forced to their assigned Tenant. Existing instance Tenant assignments cannot be changed. Blank passwords preserve the existing password on updates and are required for creates. Metadata must be valid JSON when provided.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ImportInstancesInput' }
      responses:
        '200':
          description: Import completed or dry-run summary.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstanceImportResult' }
        '400':
          description: Import validation failed; no rows persisted.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/InstanceImportResult' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/instances/test-connectivity:
    post:
      tags: [Instances]
      summary: Test unsaved instance connectivity
      description: "Runs the same live DNS/TCP/TLS/auth/API connectivity check from create/edit modal connection values without creating or updating an enrolled instance, and without writing instance status/history. Authentication requires /v2/Auth/Login to return a 2xx/3xx response that is not a login/forbidden page and includes an OxyGen session cookie; missing/non-session cookies, forbidden pages, and login forms are auth failures. After successful authentication it probes GET /web-api/BUS/License when checkLicense is enabled, then GET /web-api/global/settings. A failed required License probe marks the result as reachable with issues and skips Settings/Triggers. checkLicense=false skips the License phase."
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/TestConnectivityInput' }
      responses:
        '200':
          description: Connectivity result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectivityResult' }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/instances/{instanceId}:
    parameters:
      - $ref: '#/components/parameters/instanceId'
    get:
      tags: [Instances]
      summary: Get enrolled OxyGen instance detail
      description: Returns a single instance scoped to the signed-in user's effective instance access; used by the per-instance dashboard.
      responses:
        '200':
          description: Scoped instance detail.
          content:
            application/json:
              schema:
                type: object
                required: [instance]
                properties:
                  instance: { $ref: '#/components/schemas/OxyGenInstance' }
        '401': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    patch:
      tags: [Instances]
      summary: Update enrolled OxyGen instance
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateInstanceInput' }
      responses:
        '200':
          description: Updated instance.
          content:
            application/json:
              schema:
                type: object
                required: [instance]
                properties:
                  instance: { $ref: '#/components/schemas/OxyGenInstance' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [Instances]
      summary: Delete enrolled OxyGen instance
      responses:
        '204': { description: Instance deleted. }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/instances/{instanceId}/test-connectivity:
    parameters:
      - $ref: '#/components/parameters/instanceId'
    post:
      tags: [Instances]
      summary: Test instance connectivity
      description: "Runs a live, CMS-initiated Phase 1 connectivity test: DNS resolution, TCP connect, HTTPS certificate inspection when applicable, OxyGen form authentication requiring /v2/Auth/Login to return a 2xx/3xx response that is not a login/forbidden page and includes an OxyGen session cookie, GET /web-api/BUS/License when check_license is enabled, then GET /web-api/global/settings. Missing/non-session cookies, forbidden pages, login forms, and inaccessible authentication skip License, Settings, and Triggers phase details. Failed required License probes are stored as license errors and skip Settings/Triggers while preserving core availability as reachable. Updates instance availability/SSL/license status and appends connectivity/license history when using the MySQL repository."
      responses:
        '200':
          description: Connectivity result.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ConnectivityResult' }
        '400': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque-session-token
  parameters:
...
    instanceId:
      name: instanceId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    Error:
      description: Error response.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
        details: {}
...
    CreateInstanceInput:
      type: object
      required: [name, username, password]
      properties:
        id: { type: string, format: uuid, description: Optional instance GUID. Used by CSV import to create instances with an exported `instance_guid`. }
        name: { type: string, minLength: 1 }
        description: { type: [string, 'null'] }
        tenantId: { type: [string, 'null'], format: uuid }
        protocol: { $ref: '#/components/schemas/InstanceProtocol' }
        host: { type: string, description: Hostname/domain without protocol, or normalized from hostname. }
        port: { type: [integer, 'null'], minimum: 1, maximum: 65535 }
        hostname: { type: string, description: Legacy/full hostname input; host or hostname is required. }
        username: { type: string, default: admin }
        password: { type: string, writeOnly: true, description: Remote OxyGen credential. Stored encrypted at rest using AES-256-GCM and never returned by API responses. Requires OXYGEN_CMS_ENCRYPTION_KEY when persisted to MySQL. }
        pollingIntervalSeconds: { type: integer, minimum: 60, maximum: 86400, default: 300 }
        isEnabled: { type: boolean, default: true }
        checkLicense: { type: boolean, default: true, description: Whether connectivity polling should call the OxyGen license/settings API. When false, the license probe is skipped. }
        archived: { type: boolean, default: false, description: Hide this server from the default instance list while retaining data/history. }
        metadata: { type: [object, 'null'], additionalProperties: true, description: Custom instance metadata shown as a JSON card on the Instance Dashboard. }
        notes: { type: [string, 'null'], description: Large instance notes block. The UI auto-detects HTML, Markdown, RTF, or plain text for dashboard display. }
    TestConnectivityInput:
      type: object
      required: [username, password]
      properties:
        protocol: { $ref: '#/components/schemas/InstanceProtocol' }
        host: { type: string, description: Hostname/domain without protocol, or normalized from hostname. }
        port: { type: [integer, 'null'], minimum: 1, maximum: 65535 }
        hostname: { type: string, description: Legacy/full hostname input; host or hostname is required. }
        username: { type: string, default: admin }
        password: { type: string, writeOnly: true, description: Remote OxyGen credential used only for this live test and never persisted by this endpoint. }
        checkLicense: { type: boolean, default: true, description: Whether this live test should call the OxyGen License phase before Settings. When false, the License phase is skipped. }
    UpdateInstanceInput:
      allOf:
        - $ref: '#/components/schemas/CreateInstanceInput'
        - type: object
          properties:
            password: { type: string, writeOnly: true, description: Blank or omitted means unchanged. Non-blank replacement credentials are stored encrypted at rest and never returned by API responses. }
    ImportInstancesInput:
      type: object
      required: [csv]
      properties:
        csv:
          type: string
          description: CSV text to import. Headers are `instance_guid,name,description,tenant,protocol,host,port,username,polling_interval_seconds,is_enabled,check_license,archived,metadata,notes,password` for global users and the same list without `tenant` for tenant-scoped users. Column order is header-driven, and trailing blank spreadsheet columns are ignored. Global create rows create missing Tenant names automatically. `metadata` must be valid JSON when nonblank; `notes` may contain HTML, Markdown, RTF, or plain text.
        dryRun:
          type: boolean
          default: false
          description: Validate and summarize without persisting changes.
    InstanceImportRowResult:
      type: object
      required: [rowNumber, instanceGuid, name, action, errors, warnings]
      properties:
        rowNumber: { type: integer, minimum: 1 }
        instanceGuid: { type: [string, 'null'], format: uuid }
        name: { type: [string, 'null'] }
        action: { type: string, enum: [create, update, skip, error] }
        errors:
          type: array
          items: { type: string }
        warnings:
          type: array
          items: { type: string }
        instance:
          oneOf:
            - $ref: '#/components/schemas/OxyGenInstance'
            - type: 'null'
    InstanceImportResult:
      type: object
      required: [dryRun, created, updated, failed, rows]
      properties:
        dryRun: { type: boolean }
        created: { type: integer, minimum: 0 }
        updated: { type: integer, minimum: 0 }
        failed: { type: integer, minimum: 0 }
        rows:
          type: array
          items: { $ref: '#/components/schemas/InstanceImportRowResult' }
...
    OxyGenInstance:
      type: object
      required: [id, name, description, tenantId, protocol, host, port, hostname, baseUrl, launchUrl, apiBaseUrl, username, pollingIntervalSeconds, isEnabled, checkLicense, archived, metadata, notes, status, createdAt, updatedAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: [string, 'null'] }
        tenantId: { type: [string, 'null'], format: uuid }
        protocol: { $ref: '#/components/schemas/InstanceProtocol' }
        host: { type: string }
        port: { type: [integer, 'null'] }
        hostname: { type: string }
        baseUrl: { type: string, format: uri }
        launchUrl: { type: string, format: uri }
        apiBaseUrl: { type: string, format: uri }
        username: { type: string }
        pollingIntervalSeconds: { type: integer }
        isEnabled: { type: boolean }
        checkLicense: { type: boolean, description: Whether the OxyGen license/settings API should be checked during polling. }
        archived: { type: boolean, description: Hidden from default instance lists while data is retained. }
        metadata: { type: [object, 'null'], additionalProperties: true }
        notes: { type: [string, 'null'] }
        status: { $ref: '#/components/schemas/InstanceStatus' }
        sslValid: { type: [boolean, 'null'] }
        sslExpiresAt: { type: [string, 'null'], format: date-time }
        lastCheckedAt: { type: [string, 'null'], format: date-time }
        lastSuccessAt: { type: [string, 'null'], format: date-time }
        lastFailureAt: { type: [string, 'null'], format: date-time }
        uptimePercent24h: { type: [number, 'null'] }
        uptimePercent7d: { type: [number, 'null'] }
        responseTimeMs: { type: [integer, 'null'] }
        lastError: { type: [string, 'null'] }
        processingStatus: { $ref: '#/components/schemas/ComponentStatus' }
        emmQueueStatus: { $ref: '#/components/schemas/ComponentStatus' }
        smsStatus: { $ref: '#/components/schemas/ComponentStatus' }
        hangfireStatus: { $ref: '#/components/schemas/ComponentStatus' }
        licenseKey: { type: [string, 'null'] }
        licenseStatus: { $ref: '#/components/schemas/LicenseStatus' }
        licenseJson: { description: Raw payload returned by the OxyGen license probe for the License JSON dialog. Skipped License phases caused by connection/auth blockers preserve the prior license rollup and do not count as dashboard License Errors. }
        settingsJson: { description: Raw payload returned by GET /web-api/global/settings. The Instance Dashboard Settings card extracts non-queue BUS_Auto_Purge, OxyGen_Version, and ClientDomain values from this payload; queue enable/paused values from BUS/EMM/SMS/Hangfire groups render on Workflow & Components rows; the Settings dialog shows the full JSON. }
        workflowSummaryJson: {}
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    ConnectivityStepResult:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
        skipped: { type: boolean, nullable: true }
        message: { type: string, nullable: true }
        httpStatusCode: { type: integer, nullable: true }
        durationMs: { type: integer, minimum: 0, nullable: true, description: Step-level phase timing when captured. }
        errorCode: { type: string, nullable: true }
        valid: { type: boolean, nullable: true }
        expiresAt: { type: string, format: date-time, nullable: true }
        address: { type: string, nullable: true, description: Resolved IP address for DNS phases. }
        family: { type: integer, nullable: true, description: IP address family returned by DNS resolution. }
        host: { type: string, nullable: true, description: Host or resolved IP used by the TCP connect phase. }
        port: { type: integer, nullable: true, description: TCP port used by the connect phase. }
    LicenseProbeResult:
      type: object
      required: [step, status, key, payload]
      properties:
        step: { $ref: '#/components/schemas/ConnectivityStepResult' }
        status: { $ref: '#/components/schemas/LicenseStatus' }
        key: { type: [string, 'null'] }
        payload: {}
    ConnectivityResult:
      type: object
      required: [ok, status, message, checkedAt, durationMs, responseTimeMs, httpStatusCode, dns, connect, ssl, authentication, api, settingsJson, license]
      properties:
        ok: { type: boolean }
        status: { type: string, enum: [reachable, unreachable, auth-error, ssl-error] }
        message: { type: string }
        checkedAt: { type: string, format: date-time }
        durationMs: { type: integer, minimum: 0, description: Full connectivity refresh duration including authentication and API/license probes. }
        responseTimeMs: { type: integer, minimum: 0, nullable: true, description: Dashboard response time as Resolve + Connect + SSL + Auth for collected, non-skipped phases. }
        httpStatusCode: { type: integer, nullable: true }
        dns: { $ref: '#/components/schemas/ConnectivityStepResult' }
        connect: { $ref: '#/components/schemas/ConnectivityStepResult' }
        ssl: { $ref: '#/components/schemas/ConnectivityStepResult' }
        authentication: { $ref: '#/components/schemas/ConnectivityStepResult' }
        api: { $ref: '#/components/schemas/ConnectivityStepResult' }
        settingsJson: { description: Raw JSON payload returned by the Settings phase (`GET /web-api/global/settings`) when collected; null when Settings was skipped or failed before a JSON payload was available. }
        license: { $ref: '#/components/schemas/LicenseProbeResult' }
...
    InstanceProtocol:
      type: string
      enum: [http, https]
      default: https
    InstanceStatus:
      type: string
      enum: [unknown, up, down, auth-error, ssl-error]
    ComponentStatus:
      type: string
      enum: [unknown, ok, warning, error]
    LicenseStatus:
      type: string
      enum: [unknown, valid, expired, warning, error]
...